move lapcounts into their own cmd
